searchPokemon

+++ src/app/pokemonapp/pokemon/favorites/favorites.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { RootObject } from '../home/model';
 import { PokemonsService } from 'src/app/shared/pokemons.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-favorites',
@@ -22,7 +23,24 @@ export class FavoritesComponent implements OnInit {
       this.favoritePokemons = obj
 
     })
-    this.pokeAction.deleteFavoritePokemon(pokemon)
+
+    Swal.fire({
+      title: 'Delete to favorites',
+      text: "¿Do you want to delete this pokemon from favorites?",
+      showDenyButton: true,
+      icon: "warning",
+      confirmButtonText: 'Yes',
+      denyButtonText: `No`,
+  }).then((result) => {
+      if (result.isConfirmed) {
+        this.pokeAction.deleteFavoritePokemon(pokemon)
+        Swal.fire({ text: "pokemon has been deleted from favorites", icon: "success" })
+
+      } else {
+
+      }
+  })
+
   }
 
 }
